[MNG-6633] - Reduce memory usage of excludes
ExcludesArtifactFilter was highly inefficient.
It took the group and artifact ID of an Exclusion, concatenated them into
a new String, which was kept in memory for the whole duration
of the build and then compared that String against the concatenation
of group and artifact ID of each incoming artifact, adding more
CPU cycles than necessary.

Instead it now just takes the group and artifact ID from the Exclusion
object and compares them against the group and artifact ID of the Artifact.

The public API was kept the same, although it would be really worth
overhauling. E.g. things like Exclude extends Include violate all kinds
of design principles.